+++ axum/src/extract/path/mod.rs
@@ -14,7 +14,6 @@ use http::{request::Parts, StatusCode};
 use serde::de::DeserializeOwned;
 use std::{
     fmt,
-    ops::{Deref, DerefMut},
     sync::Arc,
 };
 
@@ -152,21 +151,7 @@ use std::{
 #[derive(Debug)]
 pub struct Path<T>(pub T);
 
-impl<T> Deref for Path<T> {
-    type Target = T;
-
-    #[inline]
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<T> DerefMut for Path<T> {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
+axum_core::__impl_deref!(Path);
 
 #[async_trait]
 impl<T, S> FromRequestParts<S> for Path<T>